server/requirements.txt
abstract==2022.4.20
asgiref==3.5.0
base32hex==1.0.2
cachetools==5.0.0
certifi==2021.10.8
charset-normalizer==2.0.12
colorama==0.4.4
colouration==2020.7.26
coreapi==2.3.3
coreschema==0.0.4
Django==4.0.3
django-filter==21.1
djangorestframework==3.13.1
drf-yasg==1.20.0
environs==9.5.0
google-auth==2.6.6
graphviz==0.20
idna==3.3
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.1
loguru==0.6.0
MarkupSafe==2.1.1
marshmallow==3.15.0
null==0.6.1
packaging==21.3
Pillow==9.1.0
psycopg2==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyJWT==2.3.0
pyparsing==3.0.7
python-dotenv==0.19.2
pytz==2022.1
requests==2.27.1
rsa==4.8
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
six==1.16.0
sqlparse==0.4.2
tzdata==2021.5
uritemplate==4.1.1
urllib3==1.26.9
username==2.1.0
validation==0.8.1
win32-setctime==1.1.0